// src/enum.ts
/** 文件上传类型枚举 */
export enum FileSuffix {
  // 图形文件
  JPG = '.jpg',
  JPEG = '.jpeg',
  PNG = '.png',
  GIF = '.gif',
  SVG = '.svg',
  BMP = '.bmp',
  TIF = '.tif',
  TIFF = '.tiff',
  // 文档文件
  TXT = '.txt',
  DOC = '.doc',
  DOCX = '.docx',
  XLS = '.xls',
  XLSX = '.xlsx',
  PPT = '.ppt',
  PPTX = '.pptx',
  PDF = '.pdf',
  // 压缩文件
  RAR = '.rar',
  ZIP = '.zip',
}